Give Vault mlock syscall without being root
The container running Vault requires IPC_LOCK
capability to prevent sensitive values from being swapped
to disk. However, in order to be able to leverage this
capability the Vault process either needs to run as root,
or have the mlock syscall permission.

It is best practice to not run privileged containers,
so I'm adding this capability to the Vault executable here.

See https://www.vaultproject.io/docs/configuration/ for details.